fix: add logic to get BIC if there are explicit accounts to build wallet state
