+++ packages/shared/lib/core/wallet/actions/buildWalletState.ts
@@ -4,6 +4,7 @@ import { updateWalletPersistedDataOnActiveProfile } from '@core/profile'
 import { IPersistedWalletData } from '../interfaces/persisted-wallet-data.interface'
 import { IWalletState } from '../interfaces/wallet-state.interface'
 import { AddressConverter, getBlockIssuerAccounts } from '../utils'
+import { DEFAULT_SYNC_OPTIONS, getTotalWalletBalance } from '..'
 
 export async function buildWalletState(
     wallet: IWallet,
@@ -48,8 +49,9 @@ export async function buildWalletState(
     let depositAddress = ''
 
     try {
-        balances = await wallet.getBalance()
+        await wallet.sync(DEFAULT_SYNC_OPTIONS)
         accountOutputs = await wallet.accounts()
+        balances = await getTotalWalletBalance(wallet.id, accountOutputs)
         // check if the mainAccountId is still valid
         if (
             walletPersistedData.mainAccountId &&
