+++ packages/shared/lib/core/wallet/actions/getClient.ts
@@ -3,8 +3,6 @@ import { selectedWalletId } from '@core/wallet/stores'
 import { api } from '@core/api'
 import { get } from 'svelte/store'
 
-export function getClient(): Promise<Client> {
-    const selectedWallet = get(selectedWalletId)
-
-    return api.getClientFromWallet(selectedWallet)
+export function getClient(walletId: string = get(selectedWalletId)): Promise<Client> {
+    return api.getClientFromWallet(walletId)
 }
