+++ packages/shared/lib/core/wallet/actions/index.ts
@@ -11,7 +11,7 @@ export * from './sendOutput'
 export * from './setOutgoingAsyncActivitiesToClaimed'
 export * from './mintNft'
 export * from './createNativeToken'
-export * from './getBicWalletBalance'
+export * from './getBicBalance'
 export * from './getOrRequestAssetFromPersistedAssets'
 export * from './activities'
 export * from './subscribeToWalletApiEvents'
@@ -40,3 +40,4 @@ export * from './setNextSelectedWallet'
 export * from './restoreBackup'
 export * from './getOutputRewards'
 export * from './getCommitteInfo'
+export * from './getTotalWalletBalance'
